requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
httpx==0.27.2
python-dotenv==1.0.1
python-multipart==0.0.20
openai==1.58.1

